+++ source/isaaclab/isaaclab/devices/openxr/retargeters/__init__.py
@@ -5,6 +5,9 @@
 """Retargeters for mapping input device data to robot commands."""
 
 from .humanoid.fourier.gr1t2_retargeter import GR1T2Retargeter, GR1T2RetargeterCfg
+from .humanoid.unitree.g1_lower_body_retargeter import G1LowerBodyRetargeter, G1LowerBodyRetargeterCfg
+from .humanoid.unitree.g1_lower_body_standing import G1LowerBodyStandingRetargeter, G1LowerBodyStandingRetargeterCfg
+from .humanoid.unitree.g1_upper_body_retargeter import G1UpperBodyRetargeter, G1UpperBodyRetargeterCfg
 from .manipulator.gripper_retargeter import GripperRetargeter, GripperRetargeterCfg
 from .manipulator.se3_abs_retargeter import Se3AbsRetargeter, Se3AbsRetargeterCfg
 from .manipulator.se3_rel_retargeter import Se3RelRetargeter, Se3RelRetargeterCfg